+++ publish/CHANGELOG.md
@@ -1,6 +1,7 @@
 - v1.52
   - Adds a new setting to disable automatic data migration.
   - Adds a new setting to disable automatic config reload (requires restart to take effect).
+  - Fixes the default biome configuration being slightly off (most notably for Meadows).
 
 - v1.51
   - Adds terrain snap support to the `objects` fields.